lib/ansiblereview/groupvars.py: Fix group_vars and inventory path
The current code assumes that the following structure is used:
inventory/
         |_ hosts
         |_ group_vars
                     |_ group1.yml
                     |_ group2.yml

Except that ansible also allows:

inventory/
         |_ hosts
         |_ group_vars
                     |_ group1
                             |_ vars.yml
                     |_ group2
                             |_ vars.yml

This means that using directly the candate path to get the group name is
wrong as it will believe that the group name is vars.yml in the second case.
Moreover, it doesn't take into account the YAML extension.

As regards the inventory directory, just compute the path before the group_vars/
directory in the file path instead of resolving the ../../ directory.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>