lib/ansiblereview/groupvars.py: Fix variable resolution
get_group_vars() from ansible.inventory.helpers seems to return variables
defined in the inventory for every groups only. Not variables in group_vars/.
This means that the "Vars should only occur once per file" check doesnt work.
It's yet an other consequence of ansible's #23001 bug fix, as group_vars/ and
host_vars/ are now loaded by plugins.

As a solution, get variables from the inventory with group.get_vars() and
use var_loaders plugins to get group_vars/ variables. With this, the check
is working again with test from :
https://github.com/willthames/ansible-review/commit/d7445ba58f7d41fa2a7ec32f11f26c93a5d0edfa

There's still a bug with this: it doesnt handle vault passwords so Ansible
will issue and error in this case.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>