lib/ansiblereview/__init__.py: Avoid warning with directory as inventory
With the move to plugins to parse inventory(ies), Ansible may spew warnings
if the inventory given is a directory as it will try to parse all files
found in the directory.
As a hack, try some known file names to find an inventory. Maybe the different
discussions about using ansible configuration files will help, dont know.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>