lib/ansiblereview/utils/yamlindent.py: handle litteral and folded scalar
One may want to use '|' or '>' to set values for a variables, as mentionned
in the YAML specs (8.1.2. Literal Style and 8.1.3. Folded Style) so it may lead
to space errors, even if they're not.

Try to handle this by detecting ': |' or ': >'. After theses two characters,
there may be more informations but for now deals only with space indent defined
with a digit.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>