lib/ansiblereview/__init__.py: Decrypt host_vars/group_vars when creating objects
Instead of having to decrypt every now and then the host_vars/group_vars
object, just decrypt it when creating HostVars and GroupVars objects.
Now, the decrypted file would be reachable at .realpath. The .patch
member has not been changed in order to disrupt things as much as possible.

By decrypting it during object creation allows also to read the "# Standards"
header than may be found in the file and set it correctly.

By default, .realpath is set to .path.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>